backend/requirements.txt
django
django-heroku
pandas
numpy
djangorestframework